move nested sequential plate tests into separate file
